feat: プライベートマッチのルームID生成機能を新規作成し、古いファイルを削除
