--- packages/offline-backend-app/src/core/create-private-match-room-id.ts
@@ -1,30 +1,0 @@
-import * as crypto from "crypto";
-
-/** ひらがな表 */
-const kana = [
-  // prettierで一文字ずつ改行されるのを防ぐため、
-  // 行ごとにスプレッド構文で展開している
-  ...["あ", "い", "う", "え", "お"],
-  ...["か", "き", "く", "け", "こ"],
-  ...["さ", "し", "す", "せ", "そ"],
-  ...["た", "ち", "つ", "て", "と"],
-  ...["な", "に", "ぬ", "ね", "の"],
-  ...["は", "ひ", "ふ", "へ", "ほ"],
-  ...["ま", "み", "む", "め", "も"],
-  ...["や", "ゆ", "よ"],
-  ...["ら", "り", "る", "れ", "ろ"],
-  ...["わ", "を", "ん"],
-];
-
-/** ルームIDの文字数 */
-const ROOM_ID_LENGTH = 5;
-
-/**
- * プライベートマッチのルームIDを生成する
- * @return 生成されたルームID
- */
-export const createPrivateMatchRoomID = (): string =>
-  Array.from(
-    { length: ROOM_ID_LENGTH },
-    () => kana[crypto.randomInt(kana.length)],
-  ).join("");
